requirements.txt
numpy==2.1.0
pillow==10.4.0
